+++ lib/_debugger.js
@@ -5,7 +5,7 @@ const path = require('path');
 const net = require('net');
 const vm = require('vm');
 const Module = require('module');
-const repl = Module.requireRepl();
+const repl = require('repl');
 const inherits = util.inherits;
 const assert = require('assert');
 const spawn = require('child_process').spawn;
