+++ lib/internal/module.js
@@ -1,6 +1,30 @@
 'use strict';
 
-module.exports.stripBOM = stripBOM;
+module.exports = { makeRequireFunction, stripBOM };
+
+// Invoke with makeRequireFunction.call(module) where |module| is the
+// Module object to use as the context for the require() function.
+function makeRequireFunction() {
+  const Module = this.constructor;
+  const self = this;
+
+  function require(path) {
+    return self.require(path);
+  }
+
+  require.resolve = function(request) {
+    return Module._resolveFilename(request, self);
+  };
+
+  require.main = process.mainModule;
+
+  // Enable support to add extra extension types.
+  require.extensions = Module._extensions;
+
+  require.cache = Module._cache;
+
+  return require;
+}
 
 /**
  * Remove byte order marker. This catches EF BB BF (the UTF-8 BOM)
