+++ lib/module.js
@@ -274,17 +274,6 @@ Module._load = function(request, parent, isMain) {
     debug('Module._load REQUEST %s parent: %s', request, parent.id);
   }
 
-  // REPL is a special case, because it needs the real require.
-  if (request === 'internal/repl' || request === 'repl') {
-    if (Module._cache[request]) {
-      return Module._cache[request];
-    }
-    var replModule = new Module(request);
-    replModule._compile(NativeModule.getSource(request), `${request}.js`);
-    NativeModule._cache[request] = replModule;
-    return replModule.exports;
-  }
-
   var filename = Module._resolveFilename(request, parent);
 
   var cachedModule = Module._cache[filename];
@@ -376,36 +365,8 @@ var resolvedArgv;
 // the file.
 // Returns exception, if any.
 Module.prototype._compile = function(content, filename) {
-  var self = this;
   // remove shebang
   content = content.replace(shebangRe, '');
-
-  function require(path) {
-    return self.require(path);
-  }
-
-  require.resolve = function(request) {
-    return Module._resolveFilename(request, self);
-  };
-
-  Object.defineProperty(require, 'paths', { get: function() {
-    throw new Error('require.paths is removed. Use ' +
-                    'node_modules folders, or the NODE_PATH ' +
-                    'environment variable instead.');
-  }});
-
-  require.main = process.mainModule;
-
-  // Enable support to add extra extension types
-  require.extensions = Module._extensions;
-  require.registerExtension = function() {
-    throw new Error('require.registerExtension() removed. Use ' +
-                    'require.extensions instead.');
-  };
-
-  require.cache = Module._cache;
-
-  var dirname = path.dirname(filename);
 
   // create wrapper function
   var wrapper = Module.wrap(content);
@@ -431,8 +392,22 @@ Module.prototype._compile = function(content, filename) {
       global.v8debug.Debug.setBreakPoint(compiledWrapper, 0, 0);
     }
   }
-  var args = [self.exports, require, self, filename, dirname];
-  return compiledWrapper.apply(self.exports, args);
+  const dirname = path.dirname(filename);
+  const require = internalModule.makeRequireFunction.call(this);
+
+  Object.defineProperty(require, 'paths', { get: function() {
+    throw new Error('require.paths is removed. Use ' +
+                    'node_modules folders, or the NODE_PATH ' +
+                    'environment variable instead.');
+  }});
+
+  require.registerExtension = function() {
+    throw new Error('require.registerExtension() removed. Use ' +
+                    'require.extensions instead.');
+  };
+
+  const args = [this.exports, require, this, filename, dirname];
+  return compiledWrapper.apply(this.exports, args);
 };
 
 
@@ -498,10 +473,10 @@ Module._initPaths = function() {
   Module.globalPaths = modulePaths.slice(0);
 };
 
-// bootstrap repl
-Module.requireRepl = function() {
-  return Module._load('internal/repl', '.');
-};
+// TODO(bnoordhuis) Unused, remove in the future.
+Module.requireRepl = internalUtil.deprecate(function() {
+  return NativeModule.require('internal/repl');
+}, 'Module.requireRepl is deprecated.');
 
 Module._preloadModules = function(requests) {
   if (!Array.isArray(requests))
