+++ lib/repl.js
@@ -21,6 +21,7 @@
 
 'use strict';
 
+const internalModule = require('internal/module');
 const util = require('util');
 const inherits = util.inherits;
 const Stream = require('stream');
@@ -29,6 +30,7 @@ const path = require('path');
 const fs = require('fs');
 const rl = require('readline');
 const Console = require('console').Console;
+const Module = require('module');
 const domain = require('domain');
 const debug = util.debuglog('repl');
 
@@ -508,6 +510,8 @@ REPLServer.prototype.createContext = function() {
     context.global.global = context;
   }
 
+  const module = new Module('<repl>');
+  const require = internalModule.makeRequireFunction.call(module);
   context.module = module;
   context.require = require;
 
@@ -647,7 +651,7 @@ REPLServer.prototype.complete = function(line, callback) {
     completionGroupsLoaded();
   } else if (match = line.match(requireRE)) {
     // require('...<Tab>')
-    var exts = Object.keys(require.extensions);
+    const exts = Object.keys(this.context.require.extensions);
     var indexRe = new RegExp('^index(' + exts.map(regexpEscape).join('|') +
                              ')$');
