+++ test/parallel/test-repl.js
@@ -278,6 +278,10 @@ function error_test() {
       expect: 'undefined\n' + prompt_unix },
     { client: client_unix, send: '/* \'\n"\n\'"\'\n*/',
       expect: 'undefined\n' + prompt_unix },
+    // REPL should get a normal require() function, not one that allows
+    // access to internal modules without the --expose_internals flag.
+    { client: client_unix, send: 'require("internal/repl")',
+      expect: /^Error: Cannot find module 'internal\/repl'/ },
   ]);
 }
